+++ logic/other_debug_stuff/plant.h
@@ -1,7 +1,7 @@
 #ifndef  PLANT
 #define PLANT
 
-float step(int _step, int _force );
+float step(int _step, float _force );
 void init(void);
 
 #endif
